+++ include/webserver/sql/sql_connector.hpp
@@ -11,7 +11,6 @@
 
 class MySQLConnector {
 public:
-    // MySQLConnector();
     MySQLConnector(const std::string&, const std::string&, const std::string&, const std::string&, unsigned int);
     ~MySQLConnector();
