+++ src/sql/sql_connector.cpp
@@ -1,21 +1,4 @@
-#include "webserver/sql/MySQLConnector.hpp"
-
-// MySQLConnector::MySQLConnector() {
-//     try
-//     {
-//         driver_ = get_driver_instance();
-//         conn_ = driver_->connect("tcp://127.0.0.1:3306", "root", "Lx@259416");
-//         conn_->setSchema("WebServer_DB");
-//         Logger::getInstance().log("INFO", "MySQL connection successful!");
-//     }
-//     catch(sql::SQLException& e)
-//     {
-//         std::cerr << "MySQL Connector /C++ error:" << std::endl;
-//         std::cerr << "Error code: " << e.getErrorCode() << std::endl;
-//         std::cerr << "SQLState" << e.getSQLState() << std::endl;
-//         std::cerr << "Message" << e.what() << std::endl;
-//     }
-// }
+#include "webserver/sql/sql_connector.hpp"
 
 MySQLConnector::MySQLConnector(const std::string& host, const std::string& user, const std::string& passwd, const std::string& dbname, unsigned int port) {
     try
